feat(core): expose runCLI
