+++ packages/rslint/src/index.ts
@@ -17,12 +17,13 @@ export {
 } from './config/presets/index.js';
 
 // The ESLint v10-aligned programmatic Node.js API (issue #1106). This `Rslint`
-// class is the only linting surface exported from the package root; alongside
-// it the root exports just the config-authoring helpers (`defineConfig` /
-// `globalIgnores`) and the plugin presets. The low-level engine (the `lint`
-// convenience, `RSLintService`, and the Node backend) lives on internal
-// subpaths — `@rslint/core/internal` and `@rslint/core/service` — not on the
-// public root. (The browser/web-worker backend lives in `@rslint/wasm`.)
+// class and the `runCLI` wrapper are the only linting surfaces exported from
+// the package root; alongside them the root exports just the config-authoring
+// helpers (`defineConfig` / `globalIgnores`) and the plugin presets. The
+// low-level engine (the `lint` convenience, `RSLintService`, and the Node
+// backend) lives on internal subpaths — `@rslint/core/internal` and
+// `@rslint/core/service` — not on the public root. (The browser/web-worker
+// backend lives in `@rslint/wasm`.)
 export { Rslint } from './api/rslint.js';
 export type {
   RslintOptions,
@@ -31,3 +32,5 @@ export type {
   LintSuggestion,
   LintMessageFix,
 } from './api/rslint.js';
+export { runCLI } from './cli/cli.js';
+export type { RunCLIOptions } from './cli/cli.js';
